curvefish_update: int targetYaw keeps getAngle s16 extsh'd once for subf + final sth (#76) 95.33->96.29%
(cherry picked from commit 2799988f3cfff22ae00f3f35f0d82eb2c76ce22b)

+++ src/main/dll/dll_0103_curvefish.c
@@ -105,7 +105,7 @@ void curvefish_update(int obj)
     f32 dy;
     f32 dz;
     f32 mag;
-    s16 targetYaw;
+    int targetYaw;
     int yawDelta;
 
     state = ((GameObject*)obj)->extra;
